Updated SQL Database services
Have not tested since I still get an error while trying to run it. Although I think the error is just from the different configuration in the settings and it is not happy with the settings on my computer. Also still need to do the Get Cluster ID but I don't know what this is supposed to do.