Update Anonymous user's data
Goes through the selected user's Locations, Notifications and Itinerary and sets the Access Token of each to "none", then set's the Anonymous data point in the User Table for that table to True.
They will not be able to retrieve their data if they do this since they will have no data.